# python/openapi.yaml
openapi: 3.0.1
info:
  title: 간단한 소셜 미디어 애플리케이션 API
  description: |
    사용자가 게시물을 생성, 조회, 업데이트, 삭제할 수 있고, 댓글을 추가하고, 게시물에 좋아요/취소할 수 있는 기본적인 소셜 네트워킹 서비스 API입니다.
    
    ## 주요 기능
    - 게시물 CRUD 작업
    - 댓글 관리
    - 좋아요 시스템
  version: 1.0.0
  contact:
    name: Contoso 제품 팀
  
servers:
  - url: http://localhost:8080/api
    description: 로컬 개발 서버

tags:
  - name: Posts
    description: 게시물 관리 작업
  - name: Comments
    description: 댓글 관리 작업
  - name: Likes
    description: 좋아요 시스템 작업

paths:
  /posts:
    get:
      tags:
        - Posts
      summary: 게시물 목록 조회
      description: 모든 최근 게시물을 조회합니다.
      operationId: getPosts
      responses:
        '200':
          description: 성공적으로 게시물 목록을 조회했습니다.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
              example:
                - id: "1"
                  username: "user123"
                  content: "첫 번째 게시물입니다!"
                  likes: 5
                  createdAt: "2025-05-30T10:00:00Z"
                  updatedAt: "2025-05-30T10:00:00Z"
                - id: "2"
                  username: "user456"
                  content: "안녕하세요 여러분!"
                  likes: 3
                  createdAt: "2025-05-30T11:00:00Z"
                  updatedAt: "2025-05-30T11:00:00Z"
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Posts
      summary: 게시물 생성
      description: 새로운 게시물을 생성합니다.
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
            example:
              username: "user123"
              content: "새로운 게시물을 작성했습니다!"
      responses:
        '201':
          description: 게시물이 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
              example:
                id: "1"
                username: "user123"
                content: "새로운 게시물을 작성했습니다!"
                likes: 0
                createdAt: "2025-05-30T10:00:00Z"
                updatedAt: "2025-05-30T10:00:00Z"
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /posts/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        description: 게시물 ID
        schema:
          type: string
        example: "1"
    
    get:
      tags:
        - Posts
      summary: 단일 게시물 조회
      description: 특정 게시물의 상세 정보를 조회합니다.
      operationId: getPostById
      responses:
        '200':
          description: 성공적으로 게시물을 조회했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
              example:
                id: "1"
                username: "user123"
                content: "첫 번째 게시물입니다!"
                likes: 5
                createdAt: "2025-05-30T10:00:00Z"
                updatedAt: "2025-05-30T10:00:00Z"
        '404':
          description: 게시물을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      tags:
        - Posts
      summary: 게시물 업데이트
      description: 기존 게시물의 내용을 업데이트합니다.
      operationId: updatePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
            example:
              username: "user123"
              content: "업데이트된 게시물 내용입니다!"
      responses:
        '200':
          description: 게시물이 성공적으로 업데이트되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
              example:
                id: "1"
                username: "user123"
                content: "업데이트된 게시물 내용입니다!"
                likes: 5
                createdAt: "2025-05-30T10:00:00Z"
                updatedAt: "2025-05-30T12:00:00Z"
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 게시물을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Posts
      summary: 게시물 삭제
      description: 기존 게시물을 삭제합니다.
      operationId: deletePost
      responses:
        '204':
          description: 게시물이 성공적으로 삭제되었습니다.
        '404':
          description: 게시물을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /posts/{postId}/comments:
    parameters:
      - name: postId
        in: path
        required: true
        description: 게시물 ID
        schema:
          type: string
        example: "1"
    
    get:
      tags:
        - Comments
      summary: 게시물의 댓글 목록 조회
      description: 특정 게시물의 모든 댓글을 조회합니다.
      operationId: getComments
      responses:
        '200':
          description: 성공적으로 댓글 목록을 조회했습니다.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
              example:
                - id: "1"
                  postId: "1"
                  username: "user456"
                  content: "좋은 게시물이네요!"
                  createdAt: "2025-05-30T10:30:00Z"
                  updatedAt: "2025-05-30T10:30:00Z"
                - id: "2"
                  postId: "1"
                  username: "user789"
                  content: "동의합니다!"
                  createdAt: "2025-05-30T11:00:00Z"
                  updatedAt: "2025-05-30T11:00:00Z"
        '404':
          description: 게시물을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Comments
      summary: 댓글 생성
      description: 특정 게시물에 새로운 댓글을 생성합니다.
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
            example:
              username: "user456"
              content: "훌륭한 게시물입니다!"
      responses:
        '201':
          description: 댓글이 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
              example:
                id: "1"
                postId: "1"
                username: "user456"
                content: "훌륭한 게시물입니다!"
                createdAt: "2025-05-30T10:30:00Z"
                updatedAt: "2025-05-30T10:30:00Z"
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 게시물을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /posts/{postId}/comments/{commentId}:
    parameters:
      - name: postId
        in: path
        required: true
        description: 게시물 ID
        schema:
          type: string
        example: "1"
      - name: commentId
        in: path
        required: true
        description: 댓글 ID
        schema:
          type: string
        example: "1"
    
    get:
      tags:
        - Comments
      summary: 특정 댓글 조회
      description: 특정 댓글의 상세 정보를 조회합니다.
      operationId: getCommentById
      responses:
        '200':
          description: 성공적으로 댓글을 조회했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
              example:
                id: "1"
                postId: "1"
                username: "user456"
                content: "좋은 게시물이네요!"
                createdAt: "2025-05-30T10:30:00Z"
                updatedAt: "2025-05-30T10:30:00Z"
        '404':
          description: 댓글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      tags:
        - Comments
      summary: 댓글 업데이트
      description: 기존 댓글의 내용을 업데이트합니다.
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
            example:
              username: "user456"
              content: "업데이트된 댓글 내용입니다!"
      responses:
        '200':
          description: 댓글이 성공적으로 업데이트되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
              example:
                id: "1"
                postId: "1"
                username: "user456"
                content: "업데이트된 댓글 내용입니다!"
                createdAt: "2025-05-30T10:30:00Z"
                updatedAt: "2025-05-30T12:30:00Z"
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 댓글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Comments
      summary: 댓글 삭제
      description: 기존 댓글을 삭제합니다.
      operationId: deleteComment
      responses:
        '204':
          description: 댓글이 성공적으로 삭제되었습니다.
        '404':
          description: 댓글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /posts/{postId}/likes:
    parameters:
      - name: postId
        in: path
        required: true
        description: 게시물 ID
        schema:
          type: string
        example: "1"
    
    post:
      tags:
        - Likes
      summary: 게시물 좋아요
      description: 특정 게시물에 좋아요를 추가합니다.
      operationId: likePost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
            example:
              username: "user456"
      responses:
        '201':
          description: 좋아요가 성공적으로 추가되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResponse'
              example:
                postId: "1"
                username: "user456"
                liked: true
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 게시물을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Likes
      summary: 게시물 좋아요 취소
      description: 특정 게시물의 좋아요를 취소합니다.
      operationId: unlikePost
      parameters:
        - name: username
          in: query
          required: true
          description: 사용자 이름
          schema:
            type: string
          example: "user456"
      responses:
        '204':
          description: 좋아요가 성공적으로 취소되었습니다.
        '404':
          description: 게시물 또는 좋아요를 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Post:
      type: object
      required:
        - id
        - username
        - content
        - likes
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: 게시물의 고유 식별자
          example: "1"
        username:
          type: string
          description: 게시물 작성자의 사용자 이름
          example: "user123"
        content:
          type: string
          description: 게시물의 내용
          example: "첫 번째 게시물입니다!"
        likes:
          type: integer
          description: 게시물의 좋아요 수
          minimum: 0
          example: 5
        createdAt:
          type: string
          format: date-time
          description: 게시물 생성 일시
          example: "2025-05-30T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 게시물 마지막 수정 일시
          example: "2025-05-30T10:00:00Z"
    
    CreatePostRequest:
      type: object
      required:
        - username
        - content
      properties:
        username:
          type: string
          description: 게시물 작성자의 사용자 이름
          minLength: 1
          example: "user123"
        content:
          type: string
          description: 게시물의 내용
          minLength: 1
          example: "새로운 게시물을 작성했습니다!"
    
    UpdatePostRequest:
      type: object
      required:
        - username
        - content
      properties:
        username:
          type: string
          description: 게시물 작성자의 사용자 이름
          minLength: 1
          example: "user123"
        content:
          type: string
          description: 업데이트할 게시물의 내용
          minLength: 1
          example: "업데이트된 게시물 내용입니다!"
    
    Comment:
      type: object
      required:
        - id
        - postId
        - username
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: 댓글의 고유 식별자
          example: "1"
        postId:
          type: string
          description: 댓글이 속한 게시물의 ID
          example: "1"
        username:
          type: string
          description: 댓글 작성자의 사용자 이름
          example: "user456"
        content:
          type: string
          description: 댓글의 내용
          example: "좋은 게시물이네요!"
        createdAt:
          type: string
          format: date-time
          description: 댓글 생성 일시
          example: "2025-05-30T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 댓글 마지막 수정 일시
          example: "2025-05-30T10:30:00Z"
    
    CreateCommentRequest:
      type: object
      required:
        - username
        - content
      properties:
        username:
          type: string
          description: 댓글 작성자의 사용자 이름
          minLength: 1
          example: "user456"
        content:
          type: string
          description: 댓글의 내용
          minLength: 1
          example: "훌륭한 게시물입니다!"
    
    UpdateCommentRequest:
      type: object
      required:
        - username
        - content
      properties:
        username:
          type: string
          description: 댓글 작성자의 사용자 이름
          minLength: 1
          example: "user456"
        content:
          type: string
          description: 업데이트할 댓글의 내용
          minLength: 1
          example: "업데이트된 댓글 내용입니다!"
    
    LikeRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: 좋아요를 누르는 사용자 이름
          minLength: 1
          example: "user456"
    
    LikeResponse:
      type: object
      required:
        - postId
        - username
        - liked
      properties:
        postId:
          type: string
          description: 좋아요가 추가된 게시물의 ID
          example: "1"
        username:
          type: string
          description: 좋아요를 누른 사용자 이름
          example: "user456"
        liked:
          type: boolean
          description: 좋아요 상태
          example: true
    
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: 오류 메시지
          example: "요청을 처리할 수 없습니다."
        code:
          type: string
          description: 오류 코드
          example: "BAD_REQUEST"
        details:
          type: string
          description: 오류에 대한 추가 세부 정보
          example: "username 필드는 필수입니다."